+++ packages/orm/angel_migration_runner/lib/src/mysql/table.dart
@@ -14,10 +14,19 @@ abstract class MySqlGenerator {
   static String columnType(MigrationColumn column) {
     var str = column.type.name;
 
-    // Handle reference key
-    if (str.toLowerCase() == ColumnType.int.name &&
-        column.externalReferences.isNotEmpty) {
-      return 'BIGINT UNSIGNED';
+    /*
+    // Handle reference key to serial
+    if (str.toLowerCase() == ColumnType.int.name) {
+      if (column.externalReferences.isNotEmpty) {
+
+        return 'BIGINT UNSIGNED';
+      }
+    }
+    */
+
+    // Map serial to int
+    if (str.toLowerCase() == ColumnType.serial.name) {
+      return ColumnType.int.name;
     }
 
     // Map timestamp time to datetime
@@ -79,7 +88,7 @@ abstract class MySqlGenerator {
 
       // For int based primary key, apply NOT NULL
       // and AUTO_INCREMENT
-      if (column.type == ColumnType.int) {
+      if (column.type == ColumnType.int || column.type == ColumnType.serial) {
         buf.write(' NOT NULL AUTO_INCREMENT');
       }
     }