+++ packages/orm/angel_orm/lib/src/query_base.dart
@@ -1,11 +1,15 @@
 import 'dart:async';
 
+import 'package:logging/logging.dart';
+
 import 'query_executor.dart';
 import 'union.dart';
 import 'package:optional/optional.dart';
 
 /// A base class for objects that compile to SQL queries, typically within an ORM.
 abstract class QueryBase<T> {
+  final _log = Logger('QueryBase');
+
   /// Casts to perform when querying the database.
   Map<String, String> get casts => {};
 
@@ -57,8 +61,8 @@ abstract class QueryBase<T> {
   Future<List<T>> get(QueryExecutor executor) async {
     var sql = compile({});
 
-    //print('sql = $sql');
-    //print('substitutionValues = $substitutionValues');
+    _log.fine('sql = $sql');
+    _log.fine('values = $substitutionValues');
 
     return executor.query(tableName, sql, substitutionValues).then((it) {
       return deserializeList(it);