+++ packages/orm/angel_orm_mysql/test/belongs_to_test.dart
@@ -1,14 +1,19 @@
 import 'package:angel3_migration_runner/angel3_migration_runner.dart';
 import 'package:angel3_orm/angel3_orm.dart';
+import 'package:logging/logging.dart';
 import 'package:mysql_client/mysql_client.dart';
 import 'package:test/test.dart';
 import 'common.dart';
 import 'models/book.dart';
 
-import 'models/unorthodox.dart';
 import 'util.dart';
 
 void main() {
+  Logger.root.level = Level.ALL; // defaults to Level.INFO
+  Logger.root.onRecord.listen((record) {
+    print('${record.level.name}: ${record.time}: ${record.message}');
+  });
+
   late MySQLConnection conn;
   late QueryExecutor executor;
   late MigrationRunner runner;