+++ packages/orm/angel_orm_postgres/test/custom_expr_test.dart
@@ -1,5 +1,6 @@
 import 'package:angel3_migration_runner/angel3_migration_runner.dart';
 import 'package:angel3_orm/angel3_orm.dart';
+import 'package:logging/logging.dart';
 import 'package:postgres/postgres.dart';
 import 'package:test/test.dart';
 import 'common.dart';
@@ -12,6 +13,11 @@ void main() {
   late Numbers numbersModel;
 
   setUp(() async {
+    Logger.root.level = Level.ALL; // defaults to Level.INFO
+    Logger.root.onRecord.listen((record) {
+      print('${record.level.name}: ${record.time}: ${record.message}');
+    });
+
     conn = await openPgConnection();
     executor = await createExecutor(conn);
     runner =