+++ packages/orm/angel_orm_postgres/test/has_many_test.dart
@@ -1,11 +1,17 @@
 import 'package:angel3_migration_runner/angel3_migration_runner.dart';
 import 'package:angel3_orm/angel3_orm.dart';
+import 'package:logging/logging.dart';
 import 'package:postgres/postgres.dart';
 import 'package:test/test.dart';
 import 'common.dart';
 import 'models/tree.dart';
 
 void main() {
+  Logger.root.level = Level.ALL; // defaults to Level.INFO
+  Logger.root.onRecord.listen((record) {
+    print('${record.level.name}: ${record.time}: ${record.message}');
+  });
+
   late Connection conn;
   late QueryExecutor executor;
   late MigrationRunner runner;