+++ packages/orm/angel_orm_postgres/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 8.3.1
+
+* Updated test cases
+
 ## 8.3.0
 
 * Require Dart >= 3.6